feat: 通过栏目获取相关的部门同步树

+++ rjsoft-core/src/main/java/com/rongji/rjsoft/entity/content/CmsSite.java
@@ -4,11 +4,12 @@ import com.baomidou.mybatisplus.annotation.FieldFill;
 import com.baomidou.mybatisplus.annotation.IdType;
 import com.baomidou.mybatisplus.annotation.TableField;
 import com.baomidou.mybatisplus.annotation.TableId;
-import java.time.LocalDateTime;
-import java.io.Serializable;
 import lombok.Data;
 import lombok.EqualsAndHashCode;
 import lombok.experimental.Accessors;
+
+import java.io.Serializable;
+import java.time.LocalDateTime;
 
 /**
  * <p>
@@ -74,7 +75,7 @@ public class CmsSite implements Serializable {
     /**
      * 部门id
      */
-    private Integer deptId;
+    private Long deptId;
 
     /**
      * 创建者
